refactor: CarRacing 클래스는 domain에 속하지 않으므로 디렉토리 위치 변경
