+++ src/main/java/racingcar/domain/Accelerator.java
@@ -1,4 +1,4 @@
-package domain;
+package racingcar.domain;
 
 public interface Accelerator {
     int push();
