+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package domain;
+package racingcar.domain;
 
 public class Car {
     private final String name;
