+++ src/main/java/racingcar/domain/TryCount.java
@@ -1,4 +1,4 @@
-package domain;
+package racingcar.domain;
 
 public class TryCount {
     private int value;
