+++ src/main/java/racingcar/io/InputView.java
@@ -1,6 +1,6 @@
-package io;
+package racingcar.io;
 
-import io.validator.InputValidator;
+import racingcar.io.validator.InputValidator;
 
 import java.util.List;
 import java.util.Scanner;
