+++ src/main/java/racingcar/util/RandomNumberUtils.java
@@ -1,4 +1,4 @@
-package util;
+package racingcar.util;
 
 import java.util.Random;
