+++ src/test/java/CarRacingTest.java
@@ -2,13 +2,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import domain.Car;
-import domain.CarAccelerator;
-import domain.CarRacing;
-import domain.Cars;
-import domain.TryCount;
-import io.InputView;
-import io.OutputView;
+import racingcar.CarRacing;
+import racingcar.domain.Car;
+import racingcar.domain.CarAccelerator;
+import racingcar.domain.Cars;
+import racingcar.domain.TryCount;
+import racingcar.io.InputView;
+import racingcar.io.OutputView;
 import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
